requirements.txt
Django==1.11.7
django-pagedown==1.0.5
Markdown==2.6.11
markdown2==2.3.5
Pillow==5.1.0
Pygments==2.2.0
PyMySQL==0.8.0
pytz==2018.3
